core: grandpa: move environment to its own submodule
