+++ core/finality-grandpa/src/consensus_changes.rs
@@ -18,18 +18,18 @@ use std::sync::Arc;
 
 /// Consensus-related data changes tracker.
 #[derive(Clone, Debug, Encode, Decode)]
-pub struct ConsensusChanges<H, N> {
+pub(crate) struct ConsensusChanges<H, N> {
 	pending_changes: Vec<(N, H)>,
 }
 
 impl<H: Copy + PartialEq, N: Copy + Ord> ConsensusChanges<H, N> {
 	/// Create empty consensus changes.
-	pub fn empty() -> Self {
+	pub(crate) fn empty() -> Self {
 		ConsensusChanges { pending_changes: Vec::new(), }
 	}
 
 	/// Note unfinalized change of consensus-related data.
-	pub fn note_change(&mut self, at: (N, H)) {
+	pub(crate) fn note_change(&mut self, at: (N, H)) {
 		let idx = self.pending_changes
 			.binary_search_by_key(&at.0, |change| change.0)
 			.unwrap_or_else(|i| i);
@@ -38,7 +38,7 @@ impl<H: Copy + PartialEq, N: Copy + Ord> ConsensusChanges<H, N> {
 
 	/// Finalize all pending consensus changes that are finalized by given block.
 	/// Returns true if there any changes were finalized.
-	pub fn finalize<F: Fn(N) -> ::client::error::Result<Option<H>>>(
+	pub(crate) fn finalize<F: Fn(N) -> ::client::error::Result<Option<H>>>(
 		&mut self,
 		block: (N, H),
 		canonical_at_height: F,
@@ -66,4 +66,4 @@ impl<H: Copy + PartialEq, N: Copy + Ord> ConsensusChanges<H, N> {
 }
 
 /// Thread-safe consensus changes tracker reference.
-pub type SharedConsensusChanges<H, N> = Arc<parking_lot::Mutex<ConsensusChanges<H, N>>>;
+pub(crate) type SharedConsensusChanges<H, N> = Arc<parking_lot::Mutex<ConsensusChanges<H, N>>>;
